+++ CHANGES.md
@@ -1,6 +1,7 @@
 master (in development)
 =======================
 
+- Add `fancy-battery-status-update-functions`
 - Fix detection of charging battery
 
 0.1 (Jul 29, 2014)